feat(sketch): split icon command into sync and generate
