+++ packages/sketch/src/commands/icons/generate.js
@@ -5,262 +5,124 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-/* global MSSVGImporter, NSString, NSUTF8StringEncoding */
-
-import { toString } from '@carbon/icon-helpers';
-import { Artboard, Document, Rectangle, Shape } from 'sketch/dom';
+import { Document, Text, Rectangle, Group } from 'sketch/dom';
 import { command } from '../command';
-import { syncColorStyles } from '../../sharedStyles/colors';
-import { groupByKey } from '../../tools/grouping';
 import { findOrCreatePage, selectPage } from '../../tools/page';
-import { syncSymbol } from '../../tools/symbols';
+import { syncIconSymbols } from './shared';
 
-const meta = require('@carbon/icons/build-info.json');
-const metadata = require('@carbon/icons/metadata.json');
-const { icons } = metadata;
-
-function findIconByName(name) {
-  const [basename, ...variants] = name.split('--');
-  const iconEntry = icons.find(icon => {
-    return icon.name === basename;
-  });
-
-  if (!iconEntry) {
-    console.log(`Unable to find the following icon by name ${name}`);
-    return iconEntry;
-  }
-
-  if (variants.length > 0) {
-    const icon = iconEntry.variants.find(variant => {
-      return variant.name === name;
-    });
-    return {
-      ...iconEntry,
-      ...icon,
-    };
-  }
-
-  return iconEntry;
-}
+const { categories } = require('@carbon/icons/metadata.json');
 
 export function generate() {
   command('commands/icons/generate', () => {
     const document = Document.getSelectedDocument();
-    const page = selectPage(findOrCreatePage(document, 'Icons'));
-    const sharedStyles = syncColorStyles(document);
-    const [sharedStyle] = sharedStyles.filter(
-      ({ name }) => name === 'color/black'
-    );
+    // const symbols = syncIconSymbols(document);
+    const symbols = document.getSymbols();
+    const page = selectPage(findOrCreatePage(document, 'Icons Template'));
+    const groups = [];
+    let PAGE_X_OFFSET = 0;
+    let PAGE_Y_OFFSET = 0;
 
-    if (!sharedStyle) {
-      throw new Error(
-        'Unexpected error occurred, expected shared style but found none'
-      );
+    for (const category of categories) {
+      const categoryText = createTextLayer({
+        frame: new Rectangle(0, 0),
+        text: category.name,
+        fontSize: 32,
+      });
+      const group = new Group({
+        name: category.name,
+        frame: new Rectangle(PAGE_X_OFFSET, PAGE_Y_OFFSET),
+        layers: [categoryText],
+      });
+
+      const subcategoryLayers = [];
+      let GROUP_X_OFFSET = 0;
+      let GROUP_Y_OFFSET = categoryText.frame.height + 32;
+
+      for (const subcategory of category.subcategories) {
+        const iconsGroupedByBase = getIconsGroupedByBase(subcategory.members);
+        const subcategoryText = createTextLayer({
+          frame: new Rectangle(0, 0),
+          text: subcategory.name,
+          fontSize: 24,
+        });
+        const subcategoryGroup = new Group({
+          name: subcategory.name,
+          frame: new Rectangle(GROUP_X_OFFSET, GROUP_Y_OFFSET),
+          layers: [subcategoryText],
+        });
+
+        const layers = [];
+        const MARGIN = 4;
+        let ICON_X_OFFSET = 0;
+        let ICON_Y_OFFSET = subcategoryText.frame.height + 8;
+        let COLUMN_COUNT = 0;
+
+        for (const icons of iconsGroupedByBase) {
+          for (const icon of icons) {
+            console.log(COLUMN_COUNT);
+            const symbol = symbols.find(symbol => {
+              const parts = symbol.name.split('/').map(string => string.trim());
+              const [_type, _category, _subcategory, name, size] = parts;
+              return name === icon && size === '32';
+            });
+            const instance = symbol.createNewInstance();
+            instance.frame.offset(ICON_X_OFFSET, ICON_Y_OFFSET);
+
+            layers.push(instance);
+            ICON_X_OFFSET = ICON_X_OFFSET + 32 + MARGIN;
+            COLUMN_COUNT = COLUMN_COUNT + 1;
+
+            // 8 column layout
+            if (COLUMN_COUNT > 7) {
+              ICON_X_OFFSET = 0;
+              COLUMN_COUNT = 0;
+              ICON_Y_OFFSET = ICON_Y_OFFSET + 32 + MARGIN;
+            }
+          }
+        }
+
+        subcategoryGroup.layers.push(...layers);
+        subcategoryGroup.adjustToFit();
+        subcategoryLayers.push(subcategoryGroup);
+        GROUP_Y_OFFSET = GROUP_Y_OFFSET + subcategoryGroup.frame.height + 32;
+      }
+
+      group.layers.push(...subcategoryLayers);
+      group.adjustToFit();
+      PAGE_Y_OFFSET = PAGE_Y_OFFSET + group.frame.height;
+
+      groups.push(group);
     }
 
-    const icons = normalize(meta);
-    const iconNames = Object.keys(icons);
-
-    // To help with debugging, we have `start` and `end` values here to focus on
-    // specific icon ranges. You can also work on a specific icon by finding
-    // it's index and setting the value of start to the index and end to the
-    // index + 1.
-    //
-    // To find the index, you can use:
-    //   console.log(iconNames.findIndex(name === 'name-to-find')); // 50
-    // And use that value below like:
-    //  const start = 50;
-    //  const end = 51;
-    // This will allow you to focus only on the icon named 'name-to-find'
-    const start = 0;
-    const end = iconNames.length;
-
-    // We keep track of the current X and Y offsets at the top-level, each
-    // iteration of an icon set should reset the X_OFFSET and update the
-    // Y_OFFSET with the maximum size in the icon set.
-    let X_OFFSET = 0;
-    let Y_OFFSET = 0;
-    let maxSize = -Infinity;
-
-    const artboards = iconNames.slice(start, end).flatMap((name, i) => {
-      const sizes = icons[name];
-
-      X_OFFSET = 0;
-      if (i !== 0) {
-        Y_OFFSET = Y_OFFSET + maxSize + 32;
-      }
-      maxSize = -Infinity;
-
-      return sizes.map(icon => {
-        // If our icon has an original size, we will need to render it in the
-        // original size and then resize it to the appropriate artboard size
-        const size = icon.original || icon.size;
-        const descriptor = Object.assign({}, icon.descriptor);
-
-        // We push a transparent rectangle to mirror the "bounding box" found in
-        // icon artboards that is stripped by our build process. Including this
-        // makes sure that our icon renders true to the path data
-        descriptor.content.push({
-          elem: 'rect',
-          attrs: {
-            width: size,
-            height: size,
-            fill: 'none',
-          },
-        });
-
-        const layer = createSVGLayer(icon.descriptor);
-
-        layer.name = icon.basename;
-        layer.rect = {
-          origin: {
-            x: 0,
-            y: 0,
-          },
-          size: {
-            width: icon.size,
-            height: icon.size,
-          },
-        };
-
-        const info = findIconByName(name);
-        const categories = info.categories;
-        let symbolName = name;
-
-        if (sizes.length !== 1) {
-          symbolName = `${name} / ${icon.size}`;
-        }
-
-        if (Array.isArray(categories) && categories.length > 0) {
-          const [category] = categories;
-          symbolName = `${category.name} / ${category.subcategory} / ${symbolName}`;
-        }
-
-        symbolName = `icon / ${symbolName}`;
-
-        const artboard = new Artboard({
-          name: symbolName,
-          frame: new Rectangle(X_OFFSET, Y_OFFSET, icon.size, icon.size),
-          layers: [layer],
-        });
-
-        if (size > maxSize) {
-          maxSize = size;
-        }
-
-        X_OFFSET = X_OFFSET + icon.size + 8;
-
-        const [group] = artboard.layers;
-
-        // Last layer will be the transparent rectangle we added above
-        const paths = group.layers.slice(0, -1).map(layer => layer.duplicate());
-
-        // We split things out into fillPaths and innerPaths, allowing us to
-        // style them independent of each other in the symbol. Useful for
-        // two-tone icons.
-        const { fillPaths = [], innerPaths = [] } = groupByKey(
-          paths,
-          (path, i) => {
-            const node = icon.descriptor.content[i];
-            if (node.attrs['data-icon-path'] === 'inner-path') {
-              return 'innerPaths';
-            }
-            return 'fillPaths';
-          }
-        );
-
-        let shape;
-        if (fillPaths.length === 1) {
-          shape = fillPaths[0];
-          shape.name = 'Fill';
-          shape.style = sharedStyle.style;
-          shape.sharedStyleId = sharedStyle.id;
-        } else {
-          // If we have multiple fill paths, we need to consolidate them into a
-          // single Shape so that we can style the icon with one override in the
-          // symbol
-          shape = new Shape({
-            name: 'Fill',
-            frame: new Rectangle(0, 0, icon.size, icon.size),
-            layers: fillPaths,
-            style: sharedStyle.style,
-            sharedStyleId: sharedStyle.id,
-          });
-        }
-
-        shape.style.borders = [];
-
-        for (const innerPath of innerPaths) {
-          innerPath.name = 'Inner Fill';
-          innerPath.style = sharedStyle.style;
-          innerPath.style.opacity = 0;
-          innerPath.sharedStyleId = sharedStyle.id;
-        }
-
-        artboard.layers.push(shape, ...innerPaths);
-        group.remove();
-
-        return syncSymbol(document, artboard.name, {
-          name: artboard.name,
-          frame: artboard.frame,
-          layers: artboard.layers,
-          background: artboard.background,
-        });
-      });
-    });
-
-    page.layers.push(...artboards);
+    page.layers.push(...groups);
   });
 }
 
-/**
- * Normalize a collection of icons by their basename
- * @param {Array<Icon>} icons
- * @returns {object}
- */
-function normalize(icons) {
-  // Collect all icons and group them by their base names. The value of the
-  // basename key is the array of all sizes for that icon
-  const iconsByBasename = icons.reduce((acc, icon) => {
-    // Ignore glyphs
-    if (!icon.size) {
-      return acc;
-    }
-    const name = icon.basename;
-    if (acc[name]) {
+function getIconsGroupedByBase(members) {
+  return Object.values(
+    members.reduce((acc, member) => {
+      const [type] = member.split('--');
+      if (acc[type]) {
+        return {
+          ...acc,
+          [type]: acc[type].concat(member),
+        };
+      }
       return {
         ...acc,
-        [name]: acc[name].concat(icon).sort(sortBySize),
+        [type]: [member],
       };
-    }
-    return {
-      ...acc,
-      [name]: [icon],
-    };
-  }, {});
-
-  return iconsByBasename;
+    }, {})
+  );
 }
 
-function sortBySize(a, b) {
-  return b.size - a.size;
-}
-
-/**
- * Create a layer from an SVG descriptor
- *
- * Reference:
- * https://github.com/airbnb/react-sketchapp/blob/aa3070556c47883974edbc7f78978c421a8199f7/src/jsonUtils/sketchImpl/makeSvgLayer.js#L12
- *
- * @param {object} svg
- * @returns {Layer}
- */
-function createSVGLayer(svg) {
-  const svgString = NSString.stringWithString(toString(svg));
-  const svgData = svgString.dataUsingEncoding(NSUTF8StringEncoding);
-  const svgImporter = MSSVGImporter.svgImporter();
-  svgImporter.prepareToImportFromData(svgData);
-  const svgLayer = svgImporter.importAsLayer();
-
-  return svgLayer;
+function createTextLayer({ text, frame, fontSize }) {
+  return new Text({
+    frame,
+    text,
+    style: {
+      fontFamily: 'IBM Plex Sans',
+      fontSize,
+    },
+  });
 }
