+++ packages/sketch/src/commands/index.js
@@ -13,6 +13,7 @@ import '@babel/polyfill';
 // incremental rebuilds.
 export { sync as syncColors, generate as generateColors } from './colors';
 export { generate as generateIcons } from './icons';
+export { sync as syncIcons } from './icons';
 export { sync as syncThemes, generate as generateThemes } from './themes';
 export { sync as syncType, generate as generateType } from './type';
